Refactor: Remove projectsSectionMore and adjust skill categories
This commit removes the `projectsSectionMore` property from the `siteContent` interface and makes the corresponding UI element conditional. The skill categories have also been updated to reflect a more accurate representation of the skillset. The "management" category has been added to the list of available skill categories and the order of categories in `SkillsSection` has been changed. The link to the GitHub profile has also been updated to be the correct URL.